refactor(template): group modules by responsibility
