+++ ai/llm/template/integrations/__init__.py
@@ -1,0 +1,1 @@
+"""Configuration, localization and tool-registry adapters for the facade."""
