+++ test/unit/ai/llm/template/test_patches.py
@@ -1,8 +1,7 @@
 from dataclasses import replace
 
-from ai.llm.template.dialog import build_dialog_section
-from ai.llm.template.dialog_context import DialogTemplateContext
-from ai.llm.template.patches import apply_field_patches, apply_requirement_patches
+from ai.llm.template.dialog import DialogTemplateContext, build_dialog_section
+from ai.llm.template.dialog.patches import apply_field_patches, apply_requirement_patches
 from sdk.types import (
     FieldPatch,
     OutputContractPatch,
